+++ SmallBeachFarm/Framework/FarmPatcher.cs
@@ -20,6 +20,9 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework
         /// <summary>Use the beach's background music (i.e. wave sounds) on the beach farm.</summary>
         private static bool UseBeachMusic;
 
+        /// <summary>Whether to add the campfire to the farm map.</summary>
+        private static bool AddCampfire;
+
         /// <summary>Get whether the given location is the Small Beach Farm.</summary>
         private static Func<GameLocation, bool> IsSmallBeachFarm;
 
@@ -36,12 +39,14 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework
         /// <summary>Initialize the Harmony patches.</summary>
         /// <param name="harmony">The Harmony patching API.</param>
         /// <param name="monitor">Encapsulates logging for the Harmony patch.</param>
+        /// <param name="addCampfire">Whether to add the campfire to the farm map.</param>
         /// <param name="useBeachMusic">Use the beach's background music (i.e. wave sounds) on the beach farm.</param>
         /// <param name="isSmallBeachFarm">Get whether the given location is the Small Beach Farm.</param>
         /// <param name="isOceanTile">Get whether a given position is ocean water.</param>
-        public static void Hook(HarmonyInstance harmony, IMonitor monitor, bool useBeachMusic, Func<GameLocation, bool> isSmallBeachFarm, Func<Farm, int, int, bool> isOceanTile)
+        public static void Hook(HarmonyInstance harmony, IMonitor monitor, bool addCampfire, bool useBeachMusic, Func<GameLocation, bool> isSmallBeachFarm, Func<Farm, int, int, bool> isOceanTile)
         {
             FarmPatcher.Monitor = monitor;
+            FarmPatcher.AddCampfire = addCampfire;
             FarmPatcher.UseBeachMusic = useBeachMusic;
             FarmPatcher.IsSmallBeachFarm = isSmallBeachFarm;
             FarmPatcher.IsOceanTile = isOceanTile;
@@ -127,16 +132,21 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework
             if (!FarmPatcher.IsSmallBeachFarm(__instance))
                 return;
 
-            // add campfire (derived from StardewValley.Locations.Mountain:resetSharedState
+            // toggle campfire (derived from StardewValley.Locations.Mountain:resetSharedState
             Vector2 campfireTile = new Vector2(64, 22);
-            if (!__instance.objects.ContainsKey(campfireTile))
+            if (FarmPatcher.AddCampfire)
             {
-                __instance.objects.Add(campfireTile, new Torch(campfireTile, 146, true)
+                if (!__instance.objects.ContainsKey(campfireTile))
                 {
-                    IsOn = false,
-                    Fragility = SObject.fragility_Indestructable
-                });
+                    __instance.objects.Add(campfireTile, new Torch(campfireTile, 146, true)
+                    {
+                        IsOn = false,
+                        Fragility = SObject.fragility_Indestructable
+                    });
+                }
             }
+            else if (__instance.objects.TryGetValue(campfireTile, out SObject obj) && obj is Torch torch && torch.ParentSheetIndex == 146)
+                __instance.objects.Remove(campfireTile);
         }
 
         /// <summary>A method called via Harmony after <see cref="Farm.cleanupBeforePlayerExit"/>.</summary>