+++ SmallBeachFarm/release-notes.md
@@ -2,8 +2,9 @@
 
 # Release notes
 ## Upcoming release
+* Added config option for the campfire.
 * Improved compatibility with Better Water (thanks to laulajatar!).
-* Removed files for Eemie's Map Recolor. (Compatibility is now handled by that mod directly.)
+* Improved compatibility with Eemie's Map Recolor (now handled by that mod directly).
 * Fixed issue where another mod's compatibility patch for Small Beach Farm wouldn't be applied correctly.
 
 ## 1.5